http://128.196.219.68/jira/browse/SAK-10571 SAK-10571 => * change default scale for making them the same as course grade. * change column names for GB_LETTERGRADE_MAPPING and   GB_LETTERGRADE_PERCENT_MAPPING to LG_MAPPING_ID and   LGP_MAPPING_ID in hibernate mapping.
git-svn-id: https://source.sakaiproject.org/svn/gradebook/trunk@33015 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ gradebook/service/impl/src/java/org/sakaiproject/component/gradebook/GradebookFrameworkServiceImpl.java
@@ -329,17 +329,17 @@ public class GradebookFrameworkServiceImpl extends BaseHibernateManager implemen
   private Map getHardDefaultLetterMapping()
   {
   	Map gradeMap = new HashMap();
-		gradeMap.put("A+", new Double(98));
+		gradeMap.put("A+", new Double(100));
 		gradeMap.put("A", new Double(95));
 		gradeMap.put("A-", new Double(90));
-		gradeMap.put("B+", new Double(88));
-		gradeMap.put("B", new Double(85));
+		gradeMap.put("B+", new Double(87));
+		gradeMap.put("B", new Double(83));
 		gradeMap.put("B-", new Double(80));
-		gradeMap.put("C+", new Double(78));
-		gradeMap.put("C", new Double(75));
+		gradeMap.put("C+", new Double(77));
+		gradeMap.put("C", new Double(73));
 		gradeMap.put("C-", new Double(70));
-		gradeMap.put("D+", new Double(68));
-		gradeMap.put("D", new Double(65));
+		gradeMap.put("D+", new Double(67));
+		gradeMap.put("D", new Double(63));
 		gradeMap.put("D-", new Double(60));
 		gradeMap.put("F", new Double(0.0));
